+++ test/libsolidity/syntaxTests/viewPureChecker/immutable.sol
@@ -4,4 +4,18 @@ contract B {
         return x;
     }
 }
+
+contract C {
+    address immutable a = address(1);
+    bool immutable b = true;
+    bytes32 immutable c = keccak256("c");
+
+    function f() public pure returns (address, bool, bytes32) {
+        return (a, b, c);
+    }
+
+    function g() public pure returns (address) {
+        return C.a;
+    }
+}
 // ----