+++ Changelog.md
@@ -16,6 +16,8 @@ Compiler Features:
 Bugfixes:
 * Code Generator: Fix ICE on parenthesized custom error construction in require statement.
 * Commandline Interface: Report proper error instead of ICE on non-hex mixed-case address value given via `--libraries`.
+* ViewPureChecker: Correctly classify immutable reads as pure only when initialized with a compile-time constant expression
+   and not reassigned during construction.
 
 Build System:
 * Update minimum version requirement of Boost to 1.83.0 for Windows build. This matches the minimum version for other systems.